requirements.txt
torch>=2.5.0
torchvision>=0.20.0
diffusers>=0.36.0
transformers>=4.40.0
accelerate
safetensors
huggingface_hub
peft
bitsandbytes
Pillow
numpy
pytest>=7.0.0
ruff>=0.4.0
pre-commit>=3.0.0
